server/web_scraper/knowledge/requirements.txt
# RAG Knowledge Pipeline — Python dependencies
# Install with: pip install -r requirements.txt

requests>=2.31.0
PyYAML>=6.0
beautifulsoup4>=4.12.0
lxml>=5.0.0          # faster BS4 parser (optional but recommended)
